[chore/#241] 서비스 책임분리
